chore: clean up testing app
