--- packages/fast-form-generator-react/app/components/README.md
@@ -1,8 +1,0 @@
-# App components
-The components in this folder are for testing the form generator in the development app.
-
-There should be:
-- A React component
-- A schema conforming to the data required by the React component
-- A spec file to generate Jest snapshots
-- An optional configuration file if testing any of the optional configurations for the form generator
