--- packages/fast-form-generator-react/app/components/align-horizontal/align-horizontal.config.ts
@@ -1,5 +1,0 @@
-import { FormComponentMappingToPropertyNamesProps } from "../../../src/form/form.props";
-
-export default {
-    alignHorizontal: ["alignHorizontal"],
-} as FormComponentMappingToPropertyNamesProps;
