--- packages/fast-form-generator-react/app/components/align-horizontal/align-horizontal.schema.json
@@ -1,21 +1,0 @@
-{
-    "$schema": "http://json-schema.org/schema#",
-    "title": "Component with align horizontal",
-    "description": "A test component's schema definition.",
-    "type": "object",
-    "id": "alignHorizontal",
-    "properties": {
-        "alignHorizontal": {
-            "title": "Align horizontal",
-            "type": "string",
-            "enum": [
-                "left",
-                "center",
-                "right"
-            ]
-        }
-    },
-    "required": [
-        "alignHorizontal"
-    ]
-}
