--- packages/fast-form-generator-react/app/components/align-horizontal/align-horizontal.tsx
@@ -1,20 +1,0 @@
-import * as React from "react";
-
-export enum AlignHorizontalPositions {
-    left = "left",
-    right = "right",
-}
-
-export interface AlignHorizontalProps {
-    alignHorizontal: AlignHorizontalPositions;
-}
-
-/**
- * This test components API should have:
- * - a required property which maps to a configuration
- */
-export default class AlignHorizontal extends React.Component<AlignHorizontalProps, {}> {
-    public render(): JSX.Element {
-        return <span>{this.props.alignHorizontal}</span>;
-    }
-}
