--- packages/fast-form-generator-react/app/components/align-vertical/align-vertical.config.ts
@@ -1,5 +1,0 @@
-import { FormComponentMappingToPropertyNamesProps } from "../../../src/form/form.props";
-
-export default {
-    alignVertical: ["alignVertical"],
-} as FormComponentMappingToPropertyNamesProps;
