--- packages/fast-form-generator-react/app/components/align-vertical/align-vertical.schema.json
@@ -1,21 +1,0 @@
-{
-    "$schema": "http://json-schema.org/schema#",
-    "title": "Component with align vertical",
-    "description": "A test component's schema definition.",
-    "type": "object",
-    "id": "alignVertical",
-    "properties": {
-        "alignVertical": {
-            "title": "Align vertical",
-            "type": "string",
-            "enum": [
-                "top",
-                "bottom",
-                "center"
-            ]
-        }
-    },
-    "required": [
-        "alignVertical"
-    ]
-}
