--- packages/fast-form-generator-react/app/components/checkbox/checkbox.schema.json
@@ -1,16 +1,0 @@
-{
-    "$schema": "http://json-schema.org/schema#",
-    "title": "Component with checkbox",
-    "description": "A test component's schema definition.",
-    "type": "object",
-    "id": "checkbox",
-    "properties": {
-        "toggle": {
-            "title": "Checkbox",
-            "type": "boolean"
-        }
-    },
-    "required": [
-        "toggle"
-    ]
-}
