--- packages/fast-form-generator-react/app/components/number-field/number-field.schema.json
@@ -1,32 +1,0 @@
-{
-    "$schema": "http://json-schema.org/schema#",
-    "title": "Component with number-field",
-    "description": "A test component's schema definition.",
-    "type": "object",
-    "id": "numberField",
-    "properties": {
-        "level": {
-            "title": "Level",
-            "type": "number",
-            "enum": [
-                1,
-                2,
-                3,
-                4
-            ],
-            "default": 1
-        },
-        "quantity": {
-            "title": "Quantity",
-            "type": "number",
-            "example": 10,
-            "minimum": 0,
-            "maximum": 50,
-            "multipleOf": 5
-        }
-    },
-    "required": [
-        "level",
-        "quantity"
-    ]
-}
