--- packages/fast-form-generator-react/app/components/number-field/number-field.tsx
@@ -1,28 +1,0 @@
-import * as React from "react";
-
-export enum Levels {
-    _1 = 1,
-    _2 = 2,
-    _3 = 3,
-    _4 = 4,
-}
-
-export interface NumberFieldProps {
-    level: Levels;
-    quantity: number;
-}
-
-/**
- * This test components API should have:
- * - a required number-field property which should display as a select dropdown because it is an enum
- * - a required number-field which should display as a text box
- */
-export default class NumberField extends React.Component<NumberFieldProps, {}> {
-    public render(): JSX.Element {
-        return (
-            <span>
-                {this.props.level} - {this.props.quantity}
-            </span>
-        );
-    }
-}
