--- packages/fast-form-generator-react/app/components/text-field/text-field.schema.json
@@ -1,27 +1,0 @@
-{
-    "$schema": "http://json-schema.org/schema#",
-    "title": "Component with text-field",
-    "description": "A test component's schema definition.",
-    "type": "object",
-    "id": "textField",
-    "properties": {
-        "tag": {
-            "title": "Tag",
-            "type": "string",
-            "enum": [
-                "span",
-                "button"
-            ],
-            "default": "button"
-        },
-        "text": {
-            "title": "Text",
-            "type": "string",
-            "example": "Example text"
-        }
-    },
-    "required": [
-        "tag",
-        "text"
-    ]
-}
