--- packages/fast-form-generator-react/app/components/theme/theme.config.ts
@@ -1,5 +1,0 @@
-import { FormComponentMappingToPropertyNamesProps } from "../../../src/form/form.props";
-
-export default {
-    theme: ["theme"],
-} as FormComponentMappingToPropertyNamesProps;
