--- packages/fast-form-generator-react/app/components/theme/theme.schema.json
@@ -1,20 +1,0 @@
-{
-    "$schema": "http://json-schema.org/schema#",
-    "title": "Component with theme",
-    "description": "A test component's schema definition.",
-    "type": "object",
-    "id": "theme",
-    "properties": {
-        "theme": {
-            "title": "Theme",
-            "type": "string",
-            "enum": [
-                "light",
-                "dark"
-            ]
-        }
-    },
-    "required": [
-        "theme"
-    ]
-}
