+++ packages/fast-form-generator-react/app/configs/README.md
@@ -1,0 +1,7 @@
+# App configs
+The files in this folder are for testing the form generator.
+
+There should be:
+- A schema conforming to a sample set of data
+- An optional React component as an `index.tsx` file
+- An optional configuration file for extra options specified in the root README.mds
