+++ packages/fast-form-generator-react/app/configs/align-horizontal/align-horizontal.config.ts
@@ -1,0 +1,5 @@
+import { FormComponentMappingToPropertyNamesProps } from "../../../src/form/form.props";
+
+export default {
+    alignHorizontal: ["alignHorizontal", "optionalAlignHorizontal"],
+} as FormComponentMappingToPropertyNamesProps;
