+++ packages/fast-form-generator-react/app/configs/align-horizontal/align-horizontal.schema.json
@@ -1,0 +1,30 @@
+{
+    "$schema": "http://json-schema.org/schema#",
+    "title": "Component with align horizontal",
+    "description": "A test component's schema definition.",
+    "type": "object",
+    "id": "alignHorizontal",
+    "properties": {
+        "alignHorizontal": {
+            "title": "Align horizontal",
+            "type": "string",
+            "enum": [
+                "left",
+                "center",
+                "right"
+            ]
+        },
+        "optionalAlignHorizontal": {
+            "title": "Optional align horizontal",
+            "type": "string",
+            "enum": [
+                "left",
+                "center",
+                "right"
+            ]
+        }
+    },
+    "required": [
+        "alignHorizontal"
+    ]
+}
