+++ packages/fast-form-generator-react/app/configs/align-vertical/align-vertical.config.ts
@@ -1,0 +1,5 @@
+import { FormComponentMappingToPropertyNamesProps } from "../../../src/form/form.props";
+
+export default {
+    alignVertical: ["alignVertical", "optionalAlignVertical"],
+} as FormComponentMappingToPropertyNamesProps;
