+++ packages/fast-form-generator-react/app/configs/align-vertical/align-vertical.schema.json
@@ -1,0 +1,30 @@
+{
+    "$schema": "http://json-schema.org/schema#",
+    "title": "Component with align vertical",
+    "description": "A test component's schema definition.",
+    "type": "object",
+    "id": "alignVertical",
+    "properties": {
+        "alignVertical": {
+            "title": "Align vertical",
+            "type": "string",
+            "enum": [
+                "top",
+                "bottom",
+                "center"
+            ]
+        },
+        "optionalAlignVertical": {
+            "title": "Optional align vertical",
+            "type": "string",
+            "enum": [
+                "top",
+                "bottom",
+                "center"
+            ]
+        }
+    },
+    "required": [
+        "alignVertical"
+    ]
+}
