+++ packages/fast-form-generator-react/app/configs/arrays/arrays.schema.json
