+++ packages/fast-form-generator-react/app/configs/attribute-assignment/attribute-assignment.config.ts
