+++ packages/fast-form-generator-react/app/configs/checkbox/checkbox.schema.json
@@ -1,0 +1,20 @@
+{
+    "$schema": "http://json-schema.org/schema#",
+    "title": "Component with checkbox",
+    "description": "A test component's schema definition.",
+    "type": "object",
+    "id": "checkbox",
+    "properties": {
+        "toggle": {
+            "title": "Required Checkbox",
+            "type": "boolean"
+        },
+        "optionalToggle": {
+            "title": "Optional Checkbox",
+            "type": "boolean"
+        }
+    },
+    "required": [
+        "toggle"
+    ]
+}
