+++ packages/fast-form-generator-react/app/configs/general/general.config.ts
