+++ packages/fast-form-generator-react/app/configs/number-field/number-field.schema.json
@@ -1,0 +1,36 @@
+{
+    "$schema": "http://json-schema.org/schema#",
+    "title": "Component with number-field",
+    "description": "A test component's schema definition.",
+    "type": "object",
+    "id": "numberField",
+    "properties": {
+        "level": {
+            "title": "Level",
+            "type": "number",
+            "enum": [
+                1,
+                2,
+                3,
+                4
+            ],
+            "default": 1
+        },
+        "quantity": {
+            "title": "Quantity",
+            "type": "number",
+            "example": 10,
+            "minimum": 0,
+            "maximum": 50,
+            "multipleOf": 5
+        },
+        "optionalNumber": {
+            "title": "Optional number field",
+            "type": "number"
+        }
+    },
+    "required": [
+        "level",
+        "quantity"
+    ]
+}
