+++ packages/fast-form-generator-react/app/configs/one-of/one-of.schema.json
