+++ packages/fast-form-generator-react/app/configs/textarea/textarea.schema.json
@@ -1,0 +1,41 @@
+{
+    "$schema": "http://json-schema.org/schema#",
+    "title": "Component with text-field",
+    "description": "A test component's schema definition.",
+    "type": "object",
+    "id": "textField",
+    "properties": {
+        "tag": {
+            "title": "Tag",
+            "type": "string",
+            "enum": [
+                "span",
+                "button"
+            ],
+            "default": "button"
+        },
+        "text": {
+            "title": "Text",
+            "type": "string",
+            "example": "Example text"
+        },
+        "optionalText": {
+            "title": "Optional text",
+            "type": "string",
+            "example": "Example text"
+        },
+        "optionalTag": {
+            "title": "Optional tag",
+            "type": "string",
+            "enum": [
+                "span",
+                "button"
+            ],
+            "default": "button"
+        }
+    },
+    "required": [
+        "tag",
+        "text"
+    ]
+}
