+++ packages/fast-form-generator-react/app/configs/theme/theme.config.ts
@@ -1,0 +1,5 @@
+import { FormComponentMappingToPropertyNamesProps } from "../../../src/form/form.props";
+
+export default {
+    theme: ["theme", "optionalTheme"],
+} as FormComponentMappingToPropertyNamesProps;
