+++ packages/fast-form-generator-react/app/configs/weight-properties/weight-properties.config.ts
