+++ packages/fast-form-generator-react/app/configs/weight-properties/weight-properties.schema.json
